changes test framwork to jest
