+++ package.json
@@ -1,11 +1,10 @@
 {
   "private": true,
-  "type":"module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "test": "ava"
+    "test": "jest"
   },
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^6.3.0",
@@ -34,6 +33,7 @@
     "styled-components": "^5.3.9"
   },
   "devDependencies": {
+    "@types/jest": "^29.5.0",
     "@types/node": "18.11.3",
     "@types/react": "18.0.21",
     "@types/react-dom": "18.0.6",
@@ -44,18 +44,10 @@
     "autoprefixer": "^10.4.12",
     "eslint": "^8.36.0",
     "postcss": "^8.4.18",
+    "jest": "^29.5.0",
+    "supertest": "^6.3.3",
     "tailwindcss": "^3.2.4",
     "typescript": "4.9.4",
-    "ava": "^5.2.0",
-    "ts-node": "^10.9.1"
-  },
-  "ava": {
-    "extensions": {
-      "ts": "module"
-    },
-    "nodeArguments": [
-      "--loader=ts-node/esm",
-      "--experimental-specifier-resolution=node"
-    ]
+    "ts-jest": "^29.1.0"
   }
 }
