+++ src/translateToSQL.ts
@@ -1,6 +1,9 @@
 import fetch from "isomorphic-unfetch";
 
-const translateToHuman = async (query, apiKey) => {
+const translateToSQL = async (query:any, apiKey:any,tableSchema = "") => {
+  const prompt = `Translate this natural language query into SQL without changing the case of the entries given by me:\n\n"${query}"\n\n${tableSchema ? `Use this table schema:\n\n${tableSchema}\n\n` : ''}SQL Query:`;
+  
+  console.log(prompt);
   const response = await fetch("https://api.openai.com/v1/completions", {
     method: "POST",
     headers: {
@@ -8,7 +11,7 @@ const translateToHuman = async (query, apiKey) => {
       Authorization: `Bearer ${apiKey}`,
     },
     body: JSON.stringify({
-      prompt: `Translate this SQL query into natural language:\n\n"${query}"\n\nNatural language query:`,
+      prompt,
       temperature: 0.5,
       max_tokens: 2048,
       n: 1,
@@ -29,4 +32,4 @@ const translateToHuman = async (query, apiKey) => {
   return data.choices[0].text.trim();
 };
 
-export default translateToHuman;
+export default translateToSQL;
