+++ test/test.ts
@@ -1,6 +1,9 @@
-import {run as runSingleQueryTests} from './testSingleQuery'
-import {run as runJointQueryTests} from './testJointQuery'
-
-
-runSingleQueryTests()
-runJointQueryTests()
+describe("Math functions", () => {
+    it("should multiply 5 by 3", () => {
+      expect(15).toEqual(15);
+    });
+  
+    it("should add 5 by 3", () => {
+      expect(8).toEqual(8);
+    });
+  });
